+++ test/lib/Elastica/Test/BulkTest.php
@@ -698,6 +698,39 @@ class BulkTest extends BaseTest
         $this->assertInstanceOf('Elastica\Bulk', $bulk->setRequestParam('key', 'value'));
     }
 
+    /**
+     * @group benchmark
+     */
+    public function testMemoryUsage()
+    {
+        $type = $this->_createIndex()->getType('test');
+
+        $data = array(
+            'text1' => 'Very long text for a string',
+            'text2' => 'But this is not very long',
+            'text3' => 'random or not random?',
+        );
+
+        $startMemory = memory_get_usage();
+
+        for ($n = 1; $n < 10; $n++) {
+            $docs = array();
+
+            for ($i = 1; $i <= 3000; $i++) {
+                $docs[] = new Document(uniqid(), $data);
+            }
+
+            $type->addDocuments($docs);
+            $docs = array();
+        }
+
+        unset($docs);
+
+        $endMemory = memory_get_usage();
+
+        $this->assertLessThan(1.3, $endMemory/$startMemory);
+    }
+
     public function udpDataProvider()
     {
         return array(